+++ src/auth0__email$_/index.d.ts
@@ -1,11 +1,12 @@
 import type { be_computed_pair_T, ReadableAtom_ } from '@ctx-core/nanostores'
 export declare const [
-	auth0__email__,
+	auth0__email$_,
 	auth0__email_,
 ]:be_computed_pair_T<
 	ReadableAtom_<string>
 >
 export {
-	auth0__email__ as auth0_email__,
-	auth0__email__ as auth0_email$_,
+	auth0__email$_ as auth0__email__,
+	auth0__email$_ as auth0_email__,
+	auth0__email$_ as auth0_email$_,
 }