+++ src/auth0__email$_/index.js
@@ -2,7 +2,7 @@ import { be_computed_pair_, computed_ } from '@ctx-core/nanostores'
 import { auth0__opened$_ } from '../auth0__opened$_/index.js'
 import { auth0__userinfo$_ } from '../auth0__userinfo$_/index.js'
 export const [
-	auth0__email__,
+	auth0__email$_,
 	auth0__email_,
 ] = be_computed_pair_(ctx=>{
 	const auth0__email_ = computed_(auth0__userinfo$_(ctx),
@@ -17,6 +17,7 @@ export const [
 	return auth0__email_
 })
 export {
-	auth0__email__ as auth0_email__,
-	auth0__email__ as auth0_email$_,
+	auth0__email$_ as auth0__email__,
+	auth0__email$_ as auth0_email__,
+	auth0__email$_ as auth0_email$_,
 }