+++ src/auth0__forgot_password__opened$_/index.js
@@ -5,8 +5,8 @@ export const [
 	auth0__forgot_password__opened_,
 ] = be_computed_pair_(ctx=>
 	computed_(auth0__opened$_(ctx),
-		$=>
-			$ === 'forgot_password'))
+		auth0__opened=>
+			auth0__opened === 'forgot_password'))
 export {
 	auth0__forgot_password__opened$_ as auth0__forgot_password__opened__,
 	auth0__forgot_password__opened$_ as auth0_forgot_password_opened__,