+++ src/auth0__forgot_password__validate/index.js
@@ -1,6 +1,6 @@
 import { auth0__email__validate } from '../auth0__email__validate/index.js'
 /**
- * @param data{import('../auth0__email__validate').auth0__email__validate__data_T}
+ * @param {import('../auth0__email__validate').auth0__email__validate__data_T}data
  * @returns {import('auth0-js').Auth0Error}
  */
 export function auth0__forgot_password__validate(