+++ src/auth0__oauth_token__POST__fetch/index.js
@@ -1,9 +1,5 @@
 import { fetch__template_pair_ } from '@ctx-core/fetch-undici'
 import { AUTH0_DOMAIN_ } from '../AUTH0_DOMAIN$_/index.js'
-/** @typedef {import('auth0').TokenResponse}TokenResponse */
-/** @typedef {import('auth0-js').Auth0Error}Auth0Error */
-/** @typedef {import('@ctx-core/jwt').JwtToken}JwtToken */
-/** @typedef {import('./index.d.ts').auth0__oauth_token__fetch__body_T}auth0__oauth_token__fetch__body_T */
 export const [
 	auth0__oauth_token__POST__fetch,
 	auth0__oauth_token__POST__fetch2