+++ src/auth0__signup__validate/index.js
@@ -3,7 +3,7 @@ import { auth0__email__validate } from '../auth0__email__validate/index.js'
 import { auth0__password_confirmation__validate } from '../auth0__password_confirmation__validate/index.js'
 /** @type {typeof import('./index.d.ts').auth0__signup__validate} */
 /**
- * @param data{import('../_types').auth0__signup_data_T}
+ * @param {import('../_types').auth0__signup_data_T}data
  * @returns {import('auth0-js').Auth0Error|null}
  */
 export function auth0__signup__validate(data) {