+++ src/auth0__token__is_valid_/index.js
@@ -1,6 +1,6 @@
 import { auth0__jwt__expiration__is_valid_ } from '../auth0__jwt__expiration__is_valid_/index.js'
 /**
- * @param auth0__token{import('../_types').auth0__token_T}
+ * @param {import('../_types').auth0__token_T}auth0__token
  * @returns {boolean}
  */
 export function auth0__token__is_valid_(